Move some code into seperate build.go file
